third_party/blink/web_tests/external/wpt/css/css-typed-om/stylevalue-normalization/normalize-tokens.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Normalizing "var(--A)" on a shorthand returns correct CSSUnparsedValue
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Normalizing "var(--A, 1em)" on a shorthand returns correct CSSUnparsedValue
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Normalizing "var(--A, var(--B))" on a shorthand returns correct CSSUnparsedValue
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Normalizing "calc(42px + var(--foo, 15em) + var(--bar, var(--far) + 15px))" on a shorthand returns correct CSSUnparsedValue
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
Harness: the test ran to completion.
